Extract transcript playhead follow into a dedicated hook
Move auto-scroll, unfollow-on-scroll, and Follow control state out of
TranscriptPanel into useTranscriptPlayheadFollow.
