+++ components/TranscriptPanel.tsx
@@ -38,6 +38,7 @@ import {
   mapSplitsToWords,
 } from "@/lib/edits";
 import { useTranscriptSelection } from "@/hooks/useTranscriptSelection";
+import { useTranscriptPlayheadFollow } from "@/hooks/useTranscriptPlayheadFollow";
 import { useCutRanges } from "@/hooks/useCutRanges";
 import { findActiveWordId, groupWordsBySpeaker } from "@/lib/transcript";
 
@@ -145,15 +146,6 @@ export default function TranscriptPanel() {
   const containerRef = useRef<HTMLDivElement>(null);
   const scrollRef = useRef<HTMLDivElement>(null);
   const importInputRef = useRef<HTMLInputElement>(null);
-  // Playhead follow: auto-scroll while true; manual scroll turns it off.
-  const followPlayheadRef = useRef(true);
-  /** Set by wheel/touch/rail before the matching scroll event. */
-  const userScrollGestureRef = useRef(false);
-  const [followPlayhead, setFollowPlayhead] = useState(true);
-  /** Playhead outside the readable viewport — gates the Follow control. */
-  const [playheadOffscreen, setPlayheadOffscreen] = useState(false);
-  /** Which way to jump to the playhead when unfollowed. */
-  const [followDirection, setFollowDirection] = useState<"up" | "down">("down");
   const [correcting, setCorrecting] = useState<{
     ids: number[];
     top: number;
@@ -182,58 +174,17 @@ export default function TranscriptPanel() {
     freezeSelectionRef,
   });
 
-  /** Measure whether the active word sits above/below the readable pane. */
-  const measurePlayheadAnchor = useCallback(() => {
-    const scroller = scrollRef.current;
-    if (!scroller) return { offscreen: false as const };
-    const { words, currentTime } = useEditorStore.getState();
-    const activeId = findActiveWordId(words, currentTime);
-    if (activeId < 0) {
-      setPlayheadOffscreen(false);
-      return { offscreen: false as const };
-    }
-    const el = containerRef.current?.querySelector(`[data-wid="${activeId}"]`);
-    if (!el) {
-      setPlayheadOffscreen(false);
-      return { offscreen: false as const };
-    }
-    const wordRect = el.getBoundingClientRect();
-    const scrollerRect = scroller.getBoundingClientRect();
-    // Header overlays the top of the scroller (pt-10 / scroll-pt-10).
-    const viewTop = scrollerRect.top + 40;
-    const viewBottom = scrollerRect.bottom;
-    if (wordRect.bottom < viewTop) {
-      setPlayheadOffscreen(true);
-      setFollowDirection("up");
-      return { offscreen: true as const, direction: "up" as const };
-    }
-    if (wordRect.top > viewBottom) {
-      setPlayheadOffscreen(true);
-      setFollowDirection("down");
-      return { offscreen: true as const, direction: "down" as const };
-    }
-    setPlayheadOffscreen(false);
-    return { offscreen: false as const };
-  }, []);
-
-  /** Rail / gesture hint — actual unfollow waits until scroll leaves the word. */
-  const userScrollGestureTimerRef = useRef(0);
-  const markUserScrollGesture = useCallback(() => {
-    const scroller = scrollRef.current;
-    if (!scroller || scroller.scrollHeight <= scroller.clientHeight + 1) return;
-    userScrollGestureRef.current = true;
-    // Keep the gesture armed across momentum / rail-drag scroll bursts.
-    window.clearTimeout(userScrollGestureTimerRef.current);
-    userScrollGestureTimerRef.current = window.setTimeout(() => {
-      userScrollGestureRef.current = false;
-    }, 150);
-  }, []);
-
-  const resumeFollowPlayhead = useCallback(() => {
-    followPlayheadRef.current = true;
-    setFollowPlayhead(true);
-    setPlayheadOffscreen(false);
-  }, []);
+  const {
+    showFollowControl,
+    followDirection,
+    resumeFollowPlayhead,
+    markUserScrollGesture,
+  } = useTranscriptPlayheadFollow({
+    scrollRef,
+    containerRef,
+    playing,
+    activeWordId,
+  });
 
   // Clicking a word seeks — resume following so playback stays in view.
   const onWordClick = useCallback(
@@ -388,54 +339,6 @@ export default function TranscriptPanel() {
     document.addEventListener("keydown", handler);
     return () => document.removeEventListener("keydown", handler);
   }, [selection, assigningSpeaker, correcting, openSpeakerAssign]);
-
-  // User scroll pauses follow only once the active word actually leaves view.
-  // Wheel over a non-scrollable pane is ignored (no Follow chip).
-  useEffect(() => {
-    const scroller = scrollRef.current;
-    if (!scroller) return;
-    const onGesture = () => markUserScrollGesture();
-    const onScroll = () => {
-      if (!playing) return;
-      const { offscreen } = measurePlayheadAnchor();
-      if (userScrollGestureRef.current && offscreen) {
-        followPlayheadRef.current = false;
-        setFollowPlayhead(false);
-        return;
-      }
-      // Scrolled back onto the playhead (or programmatic nearest kept it in view).
-      if (!offscreen && !followPlayheadRef.current) resumeFollowPlayhead();
-    };
-    scroller.addEventListener("wheel", onGesture, { passive: true });
-    scroller.addEventListener("touchmove", onGesture, { passive: true });
-    scroller.addEventListener("scroll", onScroll, { passive: true });
-    return () => {
-      scroller.removeEventListener("wheel", onGesture);
-      scroller.removeEventListener("touchmove", onGesture);
-      scroller.removeEventListener("scroll", onScroll);
-      window.clearTimeout(userScrollGestureTimerRef.current);
-    };
-  }, [markUserScrollGesture, measurePlayheadAnchor, resumeFollowPlayhead, playing]);
-
-  // Keep the active word in view during playback while following.
-  useEffect(() => {
-    if (!playing || !followPlayhead || activeWordId < 0) return;
-    const el = containerRef.current?.querySelector(`[data-wid="${activeWordId}"]`);
-    el?.scrollIntoView({ block: "nearest" });
-  }, [activeWordId, playing, followPlayhead]);
-
-  // Refresh offscreen/direction as the playhead advances while unfollowed.
-  useEffect(() => {
-    if (followPlayhead || !playing || activeWordId < 0) return;
-    const { offscreen } = measurePlayheadAnchor();
-    if (!offscreen) resumeFollowPlayhead();
-  }, [
-    followPlayhead,
-    playing,
-    activeWordId,
-    measurePlayheadAnchor,
-    resumeFollowPlayhead,
-  ]);
 
   const busy = status === "preparing" || status === "transcribing";
 
@@ -688,7 +591,7 @@ export default function TranscriptPanel() {
       </div>
       {/* Gradient overlay — must match the transcript panel surface */}
       <div className="absolute z-10 pointer-events-none inset-x-0 bottom-0 w-full h-20 bg-gradient-to-t from-white to-transparent dark:from-zinc-900" />
-      {!followPlayhead && playheadOffscreen && playing && activeWordId >= 0 && (
+      {showFollowControl && (
         <button
           type="button"
           onClick={resumeFollowPlayhead}
